requirements.txt
beautifulsoup4==4.12.2
blenderproc
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
h5py==3.8.0
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
progressbar==2.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytransform3d==3.1.0
PyYAML==6.0
requests==2.30.0
scipy==1.10.1